Add ASF-USB library
It contains USB HID descriptor related definitions and handlers made according to ASF library.
This should allow for developing a uniform protocol for all ASF-compatible devices (mostly XMEGA and SAM and UC3).

Luckily, Massdrop already introduced a protocol for a SAM device, which relied on ASF
and int turn has allowed to reuse a big chunk of code.